Config native discovery

+++ ntreg/src/registry_value.rs
@@ -98,7 +98,7 @@ impl Iterator for RegistryValues {
             )
         };
 
-        if status == ERROR_NO_MORE_ITEMS as i32 {
+        if status == ERROR_NO_MORE_ITEMS {
             return None;
         }
